Ses kaydı tamamlandı, gizlilik ayarları eklendi

+++ app/src/main/java/com/serkantken/ametist/activities/BaseActivity.java
@@ -10,6 +10,7 @@ import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.FieldPath;
 import com.google.firebase.firestore.FieldValue;
 import com.google.firebase.firestore.FirebaseFirestore;
+import com.orhanobut.hawk.Hawk;
 import com.serkantken.ametist.utilities.Utilities;
 
 import java.util.Date;
@@ -20,6 +21,7 @@ public class BaseActivity extends AppCompatActivity
     private DocumentReference documentReference;
     private FirebaseFirestore database;
     private FirebaseAuth auth;
+    int online, lastSeen;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -27,6 +29,11 @@ public class BaseActivity extends AppCompatActivity
         Utilities utilities = new Utilities(getApplicationContext(), this);
         database = FirebaseFirestore.getInstance();
         auth = FirebaseAuth.getInstance();
+        Hawk.init(this).build();
+        if (Hawk.contains("online"))
+            online = Hawk.get("online", 1);
+        if (Hawk.contains("lastSeen"))
+            lastSeen = Hawk.get("lastSeen", 1);
 
         documentReference = database.collection("Users").document(Objects.requireNonNull(auth.getUid()));
     }
@@ -34,14 +41,18 @@ public class BaseActivity extends AppCompatActivity
     @Override
     protected void onPause() {
         super.onPause();
-        documentReference.update("online", false);
-        documentReference.update("lastSeen", new Date().getTime());
+        if (online == 1)
+            documentReference.update("online", false);
+        if (lastSeen == 1)
+            documentReference.update("lastSeen", new Date().getTime());
     }
 
     @Override
     protected void onResume() {
         super.onResume();
-        documentReference.update("online", true);
-        documentReference.update("lastSeen", FieldValue.delete());
+        if (online == 1)
+            documentReference.update("online", true);
+        if (lastSeen == 1)
+            documentReference.update("lastSeen", FieldValue.delete());
     }
 }
